inference/requirements.txt
datasets==2.14.3
fastapi==0.104.1
gensim==4.3.2
nltk==3.8.1
numpy
pandas==2.0.3
pydantic==2.4.2
Requests==2.31.0
scikit_learn==1.3.0
scipy
sentence_transformers==2.2.2
uvicorn
streamlit
